update samba4: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@46631 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/samba4/source3/winbindd/winbindd_dsgetdcname.c
@@ -20,6 +20,7 @@
 #include "includes.h"
 #include "winbindd.h"
 #include "librpc/gen_ndr/ndr_winbind_c.h"
+#include "lib/util/string_wrappers.h"
 
 struct winbindd_dsgetdcname_state {
 	struct GUID guid;
@@ -182,7 +183,7 @@ static uint32_t get_dsgetdc_flags(uint32_t wbc_flags)
 	};
 
 	uint32_t ds_flags = 0;
-	int i = 0 ;
+	size_t i = 0;
 
 	for (i=0; i<ARRAY_SIZE(lookup_dc_flags); i++) {
 		if (wbc_flags & lookup_dc_flags[i].wbc_dc_flag) {
